Extract some code to hash_functions.cpp/hpp
